Move EOSR presenter & view to shared directory
This will be re-used for the SP journey to allow them to view a
submitted EOSR
