+++ server/routes/probationPractitionerReferrals/probationPractitionerReferralsController.ts
@@ -13,8 +13,8 @@ import SubmittedFeedbackPresenter from '../shared/appointment/feedback/submitted
 import SubmittedFeedbackView from '../shared/appointment/feedback/submittedFeedbackView'
 import ReferralCancellationReasonPresenter from './referralCancellationReasonPresenter'
 import ReferralCancellationReasonView from './referralCancellationReasonView'
-import EndOfServiceReportPresenter from '../probation-practitioner/end-of-service-report/endOfServiceReportPresenter'
-import EndOfServiceReportView from '../probation-practitioner/end-of-service-report/endOfServiceReportView'
+import EndOfServiceReportPresenter from '../shared/endOfServiceReport/endOfServiceReportPresenter'
+import EndOfServiceReportView from '../shared/endOfServiceReport/endOfServiceReportView'
 import ReferralCancellationCheckAnswersPresenter from './referralCancellationCheckAnswersPresenter'
 import ReferralCancellationCheckAnswersView from './referralCancellationCheckAnswersView'
 import { FormValidationError } from '../../utils/formValidationError'
